Compute best threading scheme

+++ build/start.py
@@ -4,6 +4,7 @@ import sys
 import os
 import copy
 import time
+import math
 import signal
 import multiprocessing
 import subprocess
@@ -21,17 +22,35 @@ PIPELINE_XML = '/tmp/pipeline-%d.xml'
 
 processes = []
 
-def msfreq(MS):
-  freqs = []
+def ms_data(MS):
+  msdata = []
   for ms in MS:
     try:
       t = table(ms + "/SPECTRAL_WINDOW")
-      freqs.append((ms, t[0]["CHAN_FREQ"][0], t[0]["CHAN_WIDTH"][0], t[0]["NUM_CHAN"]))
+      msdata.append((ms, t[0]["CHAN_FREQ"][0], t[0]["CHAN_WIDTH"][0], t[0]["NUM_CHAN"]))
     except RuntimeError:
       sys.stderr.write("Error: `%s' is an invalid MS\n" % (ms))
       sys.exit(1)
 
-  return freqs
+  return msdata
+
+def parse_subbands(subbands, maxchan):
+  bands = subbands.split(',')
+
+  n = 1000
+  for band in bands:
+    if len(band) == 1:
+      return 1
+
+    channels = band.split('-')
+    channels = map(int, channels)
+    assert(len(channels) == 2)
+    assert(channels[0] < maxchan)
+    assert(channels[1] < maxchan)
+    assert(channels[1] >= channels[0])
+    n = min(n, (channels[1] - channels[0]) + 1)
+
+  return n
 
 def sighandler(signum, frame):
   for p in processes:
@@ -43,7 +62,7 @@ if __name__ == "__main__":
   signal.signal(signal.SIGINT, sighandler)
 
   argparser = argparse.ArgumentParser(description=\
-    'Starts an imaging pipeline.')
+    'Starts an imaging pipeline. Press <Ctrl+C> to stop')
   argparser.add_argument('--threads', type=int, default=multiprocessing.cpu_count(), 
     help='max number of threads to spawn')
   argparser.add_argument('--tpldir', type=str,
@@ -54,14 +73,17 @@ if __name__ == "__main__":
     help='string of subbands with channels, e.g. "1-8,10-15,16-63"')
   argparser.add_argument('MS', metavar='MS', nargs='+',
     help='atleast one measurementset to read from')
+  argparser.add_argument('--dryrun', action='store_true', default=False,
+    help='Do not start pipeline, just show various output')
 
   cmd_args = argparser.parse_args()
 
-  server = cmd_args.tpldir + '/serverConfig.xml'
+  server   = cmd_args.tpldir + '/serverConfig.xml'
   emulator = cmd_args.tpldir + '/emulatorConfig.xml'
   pipeline = cmd_args.tpldir + '/pipelineConfig.xml'
 
-  os.makedirs(cmd_args.output)
+  if not cmd_args.dryrun:
+    os.makedirs(cmd_args.output)
   if (not os.path.isfile(server)):
     sys.exit("Error: `%s' doesn't exist" % (server))
   if (not os.path.isfile(emulator)):
@@ -70,20 +92,20 @@ if __name__ == "__main__":
     sys.exit("Error: `%s' doesn't exist" % (pipeline))
 
   # 1. Extract frequencies, channel width and number of channels
-  freqs = msfreq(cmd_args.MS)
+  msdata = ms_data(cmd_args.MS)
 
   # 2. Create the server xml and start it
   tree = ET.parse(server)
   root = tree.getroot()
   i = 0
-  for freq in freqs:
+  for data in msdata:
     streamnode = root[0][1][i]
     root[0][1].append(copy.deepcopy(streamnode))
     streamnode.set('name', 'Stream%d' % (i+1))
     streamnode[1].set('subbands', cmd_args.subbands)
-    streamnode[1].set('numChannels', str(freq[3]))
-    streamnode[1].set('frequency', str(freq[1]))
-    streamnode[1].set('width', str(freq[2]))
+    streamnode[1].set('numChannels', str(data[3]))
+    streamnode[1].set('frequency', str(data[1]))
+    streamnode[1].set('width', str(data[2]))
     streamnode[2].set('port', str(EM_START_PORT+i))
     i += 1
   root[0][1].remove(root[0][1][i])
@@ -94,15 +116,23 @@ if __name__ == "__main__":
     tree.write(f, 'utf-8')
     f.close()
 
-  time.sleep(0.1)
-  processes.append(subprocess.Popen(['./aartfaac-server', SERVER_XML]))
-  time.sleep(2)
+  if cmd_args.dryrun:
+    print "Generated %s" % (SERVER_XML)
+  else:
+    processes.append(subprocess.Popen(['./aartfaac-server', SERVER_XML]))
+    time.sleep(2)
 
   # 3. Compute number of pipelines to spawn, create the pipeline xml files and
   #    start the pipeline(s)
-  num_pipelines = cmd_args.threads - len(freqs) - 1
-  num_threads = 1
-  print "Starting %d pipelines" % (num_pipelines)
+  threads = cmd_args.threads - 1 # server thread
+  maxchan = max(msdata, key=lambda x: x[3])[3]
+  num_threads = min(threads, parse_subbands(cmd_args.subbands, maxchan))
+  num_pipelines = int(max(1, math.floor(threads/float(num_threads))))
+
+  if cmd_args.dryrun:
+    print "Starting %d pipelines using %d thread(s) each, %d total" % \
+          (num_pipelines, num_threads, num_pipelines*num_threads)
+
   for i in range(num_pipelines):
     tree = ET.parse(pipeline)
     root = tree.getroot()
@@ -118,21 +148,25 @@ if __name__ == "__main__":
       f.write('<!DOCTYPE pelican>\n\n')
       tree.write(f, 'utf-8')
       f.close()
-    time.sleep(0.1)
-    processes.append(subprocess.Popen(['./aartfaac-pipeline', filename]))
+    if cmd_args.dryrun:
+      print "Generated %s" % (filename)
+    else:
+      processes.append(subprocess.Popen(['./aartfaac-pipeline', filename]))
     
-  time.sleep(1)
+  if not cmd_args.dryrun:
+    time.sleep(1)
 
   # 4. Setup the netcat processes
   for i in range(num_pipelines):
     filename = cmd_args.output + "/pipeline-%d.log" % (i) 
-    f = open(filename, "w")
-    processes.append(subprocess.Popen(['nc', 'localhost', '%d' % (MON_START_PORT+i)], stdout=f))
-    print "Writing monitoring stats for pipeline %d to `%s'" % (i, filename)
+    if cmd_args.dryrun:
+      print "Writing monitoring stats for pipeline %d to `%s'" % (i, filename)
+    else:
+      f = open(filename, "w")
+      processes.append(subprocess.Popen(['nc', 'localhost', '%d' % (MON_START_PORT+i)], stdout=f))
   
 
   # 5. Setup xml for the emulators and start them
-  print "Starting %d emulators" % (len(freqs))
   i = 0
   for ms in cmd_args.MS:
     tree = ET.parse(emulator)
@@ -145,9 +179,12 @@ if __name__ == "__main__":
       f.write('<!DOCTYPE pelican>\n\n')
       tree.write(f, 'utf-8')
       f.close()
-    processes.append(subprocess.Popen(['./aartfaac-emulator', filename]))
+    if cmd_args.dryrun:
+      print "Generated %s" % (filename)
+    else:
+      processes.append(subprocess.Popen(['./aartfaac-emulator', filename]))
     i += 1
 
   # 6. Loop forever
-  while True:
+  while not cmd_args.dryrun:
     time.sleep(1)
